avoid building linux packages on other OSs
